### 2022-04-25 #4341 @matthiasdemuzere
Hi @crvernon,

Thanks for agreeing to be the topic editor of our work.

This work was initially submitted some time ago, see [here](https://github.com/openjournals/joss-reviews/issues/3696#issuecomment-914233486). Back then I received comments from jedbrown and kthyng. Would it make sense to ask them again?

From a content point of view, Alberto Martilli and Oscar Brousse would be valuable reviewers, although they have been providing feedback during the development phase. So they would not be 100% independent.

From the (impressively long) JOSS list, it seems to me that j3m3m1, noraloose, or alexamici could be good candidates?
