### 2022-05-03 #4341 @NoraLoose
Hi @crvernon, thanks for the invite! I think I will have to decline this time - I'm not familiar with WRF (nor with WUDAPT). Maybe we should try to recruit a frequent WRF user to the JOSS reviewer list! 😃
