### 2022-05-03 #4341 @crvernon
Thanks @NoraLoose for the quick reply! And no problem at all.  I reach out so some folks I know in the WRF community.  Have a great day!
