### 2022-05-26 #4432 @editorialbot
[REVIEW]: W2W: A Python package that injects WUDAPT’s Local Climate Zone information in WRF
**Submitting author:** @matthiasdemuzere (<a href="http://orcid.org/0000-0003-3237-4077">Matthias Demuzere</a>)
**Repository:** https://github.com/matthiasdemuzere/w2w
**Branch with paper.md** (empty if default branch):
**Version:** v0.4.1
