+++ tests/w2w_test.py
@@ -34,7 +34,7 @@ from w2w.w2w import check_lcz_integrity
 from w2w.w2w import checks_and_cleaning
 from w2w.w2w import create_lcz_extent_file
 from w2w.w2w import create_lcz_params_file
-from w2w.w2w import create_wrf_gridinfo
+#from w2w.w2w import create_wrf_gridinfo
 from w2w.w2w import expand_land_cat_parents
 from w2w.w2w import Info
 from w2w.w2w import main
@@ -356,26 +356,26 @@ def test_wrf_remove_urban_output_already_exists_is_overwritten(tmpdir, info_mock
     wrf_remove_urban(info=info, NPIX_NLC=9)
     assert m_time_old != os.path.getmtime(info.dst_nu_file)
 
-
-def test_create_wrf_gridinfo(tmpdir, info_mock):
-    info = info_mock(
-        {
-            'dst_nu_file': 'testing/5by5.nc',
-            'dst_gridinfo': os.path.join(tmpdir, 'dst_gridinfo.tif'),
-        }
-    )
-    create_wrf_gridinfo(info)
-    assert os.listdir(tmpdir) == ['dst_gridinfo.tif']
-    tif = rxr.open_rasterio(info.dst_gridinfo)
-    assert tif.rio.crs.to_proj4() == '+init=epsg:4326'
-    assert tif.rio.transform() == Affine(
-        0.01000213623046875,
-        0.0,
-        -1.4050254821777344,
-        0.0,
-        0.010000228881835938,
-        41.480000495910645,
-    )
+#TODO: replace to test new _get_wrf_grid_info() function
+# def test_create_wrf_gridinfo(tmpdir, info_mock):
+#     info = info_mock(
+#         {
+#             'dst_nu_file': 'testing/5by5.nc',
+#             'dst_gridinfo': os.path.join(tmpdir, 'dst_gridinfo.tif'),
+#         }
+#     )
+#     create_wrf_gridinfo(info)
+#     assert os.listdir(tmpdir) == ['dst_gridinfo.tif']
+#     tif = rxr.open_rasterio(info.dst_gridinfo)
+#     assert tif.rio.crs.to_proj4() == '+init=epsg:4326'
+#     assert tif.rio.transform() == Affine(
+#         0.01000213623046875,
+#         0.0,
+#         -1.4050254821777344,
+#         0.0,
+#         0.010000228881835938,
+#         41.480000495910645,
+#     )
 
 
 def test_get_SW_BW():
