### 2022-06-02 f011213
address more pre-commit checks
