+++ tests/w2w_test.py
@@ -38,8 +38,7 @@ from w2w.w2w import expand_land_cat_parents
 from w2w.w2w import Info
 from w2w.w2w import main
 from w2w.w2w import wrf_remove_urban
-
-# from w2w.w2w import create_wrf_gridinfo
+from w2w.w2w import _get_wrf_grid_info
 
 
 @pytest.fixture
@@ -466,7 +465,6 @@ def test_ucp_resampler_output_values_per_paramater(
     # Parameters should be on WRF grid size
     assert ucp_res.shape == (1, 102, 162)
     # Per parameter, check # max values and non-0 domain average
-    print('HERE!!!!', np.sum(ucp_res.data == ucp_res.data.max()))
     assert np.sum(ucp_res.data == ucp_res.data.max()) == data_sum
     assert approx(np.mean(ucp_res.data[ucp_res.data > 0])) == data_mean
     # Make sure no nans are present.
@@ -478,6 +476,8 @@ def test_ucp_resampler_output_values_per_paramater_frc_threshold(info_mock):
         {
             'src_file': 'sample_data/lcz_zaragoza.tif',
             'src_file_clean': 'testing/lcz_zaragoza_clean.tif',
+            'dst_file': 'sample_data/geo_em.d04.nc',
+            'dst_nu_file': 'testing/geo_em.d04_NoUrban.nc',
             'BUILT_LCZ': [1, 2, 3, 4, 5, 6, 7, 8, 9, 10],
         }
     )
@@ -493,7 +493,7 @@ def test_ucp_resampler_output_values_per_paramater_frc_threshold(info_mock):
     assert ucp_res.shape == (1, 102, 162)
     # Per parameter, check # max values and non-0 domain average
     assert np.sum(ucp_res.data == ucp_res.data.max()) == 1
-    assert approx(np.mean(ucp_res.data[ucp_res.data > 0])) == 0.428921
+    assert approx(np.mean(ucp_res.data[ucp_res.data > 0])) == 0.432376
     # Make sure no nans are present.
     np.sum(np.isnan(ucp_res.data)) == 0
 
@@ -504,6 +504,8 @@ def test_hgt_resampler_output_values(info_mock):
         {
             'src_file': 'sample_data/lcz_zaragoza.tif',
             'src_file_clean': 'testing/lcz_zaragoza_clean.tif',
+            'dst_file': 'sample_data/geo_em.d04.nc',
+            'dst_nu_file': 'testing/geo_em.d04_NoUrban.nc',
             'BUILT_LCZ': [1, 2, 3, 4, 5, 6, 7, 8, 9, 10],
         }
     )
@@ -520,8 +522,8 @@ def test_hgt_resampler_output_values(info_mock):
     assert ucp_res.shape == (1, 102, 162)
 
     # Check # max values and non-0 domain average
-    assert np.sum(ucp_res.data == ucp_res.data.max()) == 9
-    assert approx(np.mean(ucp_res.data[ucp_res.data > 0])) == 6.7111716
+    assert np.sum(ucp_res.data == ucp_res.data.max()) == 1
+    assert approx(np.mean(ucp_res.data[ucp_res.data > 0])) == 6.700552
 
     # Make sure no nans are present.
     assert np.sum(np.isnan(ucp_res.data)) == 0
@@ -651,6 +653,8 @@ def test_hi_resampler(info_mock):
     info = info_mock(
         {
             'src_file_clean': 'testing/lcz_zaragoza_clean.tif',
+            'dst_file': 'sample_data/geo_em.d04.nc',
[…]
