### 2022-06-02 3af355a
add test for new _get_wrf_grid_info()
