+++ w2w/w2w.py
@@ -504,19 +504,63 @@ def wrf_remove_urban(
 # Inspired by: https://fabienmaussion.info/2018/01/06/wrf-projection/
 def _get_wrf_grid_info(info: Info) -> Dict[str, Dict[str, Any]]:
 
+    # Initialize WGS84 projection
+
+    wgs_proj = pyproj.Proj(proj='latlong', datum='WGS84')
+
     # Read gridded WRF data
     dst_data = xr.open_dataset(info.dst_file)
 
-    wrf_proj = pyproj.Proj(
-        proj='lcc',  # projection type: Lambert Conformal Conic
-        lat_1=dst_data.TRUELAT1,
-        lat_2=dst_data.TRUELAT2,  # Cone intersects with the sphere
-        lat_0=dst_data.MOAD_CEN_LAT,
-        lon_0=dst_data.STAND_LON,  # Center point
-        a=6370000,
-        b=6370000,
-    )
-    wgs_proj = pyproj.Proj(proj='latlong', datum='WGS84')
+    # Projection depends on value of MAP_PROJ
+    map_proj = dst_data.MAP_PROJ
+
+    # define projections, borrowed from the salem package:
+    # shorturl.at/orEMU
+
+    # Lambert Conformal Conic
+    if map_proj == 1:
+        wrf_proj = pyproj.Proj(
+            proj='lcc',  # projection type: Lambert Conformal Conic
+            lat_1=dst_data.TRUELAT1,
+            lat_2=dst_data.TRUELAT2,  # Cone intersects with the sphere
+            lat_0=dst_data.MOAD_CEN_LAT,
+            lon_0=dst_data.STAND_LON,  # Center point
+            x_0=0,
+            y_0=0,
+            a=6370000,
+            b=6370000,
+        )
+    # Polar Stereographic
+    elif map_proj == 2:
+        wrf_proj = pyproj.Proj(
+            proj='stere',  # projection type: Lambert Conformal Conic
+            lat_ts=dst_data.TRUELAT1,
+            lat_0=90.0,
+            lon_0=dst_data.STAND_LON,  # Center point
+            x_0=0,
+            y_0=0,
+            a=6370000,
+            b=6370000,
+        )
+    # Mercator
+    elif map_proj == 3:
+        wrf_proj = pyproj.Proj(
+            proj='merc',  # projection type: Lambert Conformal Conic
+            lat_ts=dst_data.TRUELAT1,
+            lon_0=dst_data.STAND_LON,  # Center point
+            x_0=0,
+            y_0=0,
+            a=6370000,
+            b=6370000,
+        )
+
+    else:
+        message = (
+            f'{ERROR}WARNING: WRF MAP_RPOJ {map_proj} '
+            f'not implemented yet.{ENDC}'
+        )
+        print(message)
+        sys.exit(1)
 
     # Make transform
     transformer_wrf = Transformer.from_proj(wgs_proj, wrf_proj)
