+++ w2w/w2w.py
@@ -557,10 +557,10 @@ def _get_wrf_grid_info(info: Info) -> Dict[str, Dict[str, Any]]:
             lon_0=dst_data.STAND_LON,
             lat_ts=dst_data.TRUELAT1,
         )
-    # TODO: not yet OK for eqc projection
-    # Latlong - Equidistant Cylindrical
-    # Follow this: https://github.com/NCAR/wrf-python/blob/
-    # 4a9ff241c8f3615b6a5c94e10a945e8a39bdea27/src/wrf/projection.py#L928
+    #TODO: not yet OK for eqc projection
+    #Latlong - Equidistant Cylindrical
+    #Follow this: https://github.com/NCAR/wrf-python/blob/
+    #4a9ff241c8f3615b6a5c94e10a945e8a39bdea27/src/wrf/projection.py#L928
     elif map_proj == 6:
         wrf_proj = pyproj.Proj(
             proj='eqc',
@@ -581,11 +581,12 @@ def _get_wrf_grid_info(info: Info) -> Dict[str, Dict[str, Any]]:
     e, n = transformer_wrf.transform(dst_data.CEN_LON, dst_data.CEN_LAT)
 
     # Grid parameters
-    # TODO: check this -1, is done as in salem??
+    # TODO: is this correct?
+    print(">>>>>>>>>>> CHECK THIS ....")
     # https://github.com/fmaussion/salem/blob/
     # d3f2e5e340c2af36c84c82a9de6099c90fba12e8/salem/wrftools.py#L734
     dx, dy = dst_data.DX, dst_data.DY
-    nx, ny = dst_data.dims['west_east'] - 1, dst_data.dims['south_north'] - 1
+    nx, ny = dst_data.dims['west_east'], dst_data.dims['south_north']
 
     # Down left corner of the domain
     x0 = -(nx - 1) / 2.0 * dx + e
