+++ w2w/w2w.py
@@ -557,10 +557,10 @@ def _get_wrf_grid_info(info: Info) -> Dict[str, Dict[str, Any]]:
             lon_0=dst_data.STAND_LON,
             lat_ts=dst_data.TRUELAT1,
         )
-    #TODO: not yet OK for eqc projection
-    #Latlong - Equidistant Cylindrical
-    #Follow this: https://github.com/NCAR/wrf-python/blob/
-    #4a9ff241c8f3615b6a5c94e10a945e8a39bdea27/src/wrf/projection.py#L928
+    # TODO: not yet OK for eqc projection
+    # Latlong - Equidistant Cylindrical
+    # Follow this: https://github.com/NCAR/wrf-python/blob/
+    # 4a9ff241c8f3615b6a5c94e10a945e8a39bdea27/src/wrf/projection.py#L928
     elif map_proj == 6:
         wrf_proj = pyproj.Proj(
             proj='eqc',
@@ -582,7 +582,7 @@ def _get_wrf_grid_info(info: Info) -> Dict[str, Dict[str, Any]]:
 
     # Grid parameters
     # TODO: is this correct?
-    print(">>>>>>>>>>> CHECK THIS ....")
+    print('>>>>>>>>>>> CHECK THIS ....')
     # https://github.com/fmaussion/salem/blob/
     # d3f2e5e340c2af36c84c82a9de6099c90fba12e8/salem/wrftools.py#L734
     dx, dy = dst_data.DX, dst_data.DY
