### 2022-06-03 3ac3553
add wrf-python to cfg
