+++ setup.cfg
@@ -22,13 +22,13 @@ classifiers =
 [options]
 packages = find:
 install_requires =
-    wrf-python
     ecmwflibs
     numpy>=1.21
     pandas
     rasterio>=1.1.6
     rioxarray
     scipy
+    wrf-python
     xarray[io]
     importlib-metadata;python_version<"3.9"
     importlib-resources;python_version<"3.9"
