+++ setup.cfg
@@ -28,7 +28,6 @@ install_requires =
     rasterio>=1.1.6
     rioxarray
     scipy
-    wrf-python
     xarray[io]
     importlib-metadata;python_version<"3.9"
     importlib-resources;python_version<"3.9"
