### 2022-06-09 12ceb21
fix wrong reprojction to wrf domain (#60)
* fix wrong reprojction to wrf domain

* add test for new _get_wrf_grid_info()

* update results of sample files

* add different proj to _get_wrf_grid_info()

* add Latlon - cylindrical equidistant

* add test file MAP_PROJ>6

* update sample data

* test nx ny definition

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
