+++ w2w/w2w.py
@@ -6,6 +6,7 @@ import rioxarray as rxr
 import rasterio
 import xarray as xr
 from rasterio.warp import reproject, Resampling
+from rasterio.transform import Affine
 from scipy.stats import mode, truncnorm
 import os, sys
 import argparse
@@ -21,7 +22,8 @@ from typing import Tuple
 from typing import Union
 from numpy.typing import NDArray
 from scipy import stats
-from pyproj import CRS
+import pyproj
+from pyproj import CRS, Transformer
 
 if sys.version_info >= (3, 9):  # pragma: >=3.9 cover
     import importlib.metadata as importlib_metadata
@@ -152,11 +154,6 @@ def main(argv: Optional[Sequence[str]] = None) -> int:
         NPIX_NLC=args.NPIX_NLC,
     )
 
-    print(f'{FBOLD}--> Create temporary WRF grid .tif file ' f'for resampling{FEND}')
-    create_wrf_gridinfo(
-        info=info,
-    )
-
     print(f'{FBOLD}--> Create LCZ-based geo_em file{FEND}')
     nbui_max = create_lcz_params_file(
         info=info,
@@ -197,7 +194,6 @@ class Info(NamedTuple):
     src_file_clean: str
     dst_file: str
     dst_nu_file: str
-    dst_gridinfo: str
     dst_lcz_extent_file: str
     dst_lcz_params_file: str
     BUILT_LCZ: List[int]
@@ -214,10 +210,6 @@ class Info(NamedTuple):
             dst_file=os.path.join(args.io_dir, args.wrf_file),
             dst_nu_file=os.path.join(
                 args.io_dir, args.wrf_file.replace('.nc', '_NoUrban.nc')
-            ),
-            # TMP file, will be removed
-            dst_gridinfo=os.path.join(
-                args.io_dir, args.wrf_file.replace('.nc', '_gridinfo.tif')
             ),
             dst_lcz_extent_file=os.path.join(
                 args.io_dir, args.wrf_file.replace('.nc', '_LCZ_extent.nc')
@@ -508,24 +500,87 @@ def wrf_remove_urban(
     dst_data.to_netcdf(info.dst_nu_file)
 
 
-# Make WRF grid info available for Resampler (tmp file)
-def create_wrf_gridinfo(info: Info) -> None:
+# Get WRF grid info for Resampler
+# Inspired by: https://fabienmaussion.info/2018/01/06/wrf-projection/
+def _get_wrf_grid_info(info: Info) -> Dict[str, Any]:
 
-    # Read  gridded WRF data
-    dst_data = xr.open_dataset(info.dst_nu_file)
+    # Initialize WGS84 projection
+    wgs_proj = pyproj.Proj(proj='latlong', datum='WGS84')
 
-    # Create simpler WRF grid target.
-    da_lu = xr.Dataset(
-        {'LU_INDEX': (['y', 'x'], dst_data['LU_INDEX'][0, :, :].values)},
-        coords={
-            'y': dst_data.XLAT_M.values[0, :, 0],
-            'x': dst_data.XLONG_M.values[0, 0, :],
-        },
-    )
+    # Read gridded WRF data
+    dst_data = xr.open_dataset(info.dst_file)
 
-    # Add projection information as attributes, save and read back in.
-    da_lu.rio.write_crs('epsg:4326', inplace=True)
-    da_lu.rio.to_raster(info.dst_gridinfo)
+    # Projection depends on value of MAP_PROJ
+    map_proj = int(dst_data.MAP_PROJ)
+
+    # Lambert Conformal Conic
+    if map_proj == 1:
[…]
