+++ tests/w2w_test.py
@@ -23,6 +23,7 @@ from w2w.w2w import _get_lcz_arr
 from w2w.w2w import _get_lcz_band
 from w2w.w2w import _get_SW_BW
 from w2w.w2w import _get_truncated_normal_sample
+from w2w.w2w import _get_wrf_grid_info
 from w2w.w2w import _hgt_resampler
 from w2w.w2w import _hi_resampler
 from w2w.w2w import _initialize_urb_param
@@ -34,7 +35,6 @@ from w2w.w2w import check_lcz_integrity
 from w2w.w2w import checks_and_cleaning
 from w2w.w2w import create_lcz_extent_file
 from w2w.w2w import create_lcz_params_file
-from w2w.w2w import create_wrf_gridinfo
 from w2w.w2w import expand_land_cat_parents
 from w2w.w2w import Info
 from w2w.w2w import main
@@ -50,7 +50,6 @@ def info_mock():
             'src_file_clean': '',
             'dst_file': '',
             'dst_nu_file': '',
-            'dst_gridinfo': '',
             'dst_lcz_extent_file': '',
             'dst_lcz_params_file': '',
             'BUILT_LCZ': [],
@@ -74,11 +73,10 @@ def test_create_info_dict():
         built_lcz=[1, 2, 3, 4, 5, 6, 7, 8, 9, 10],
     )
     info = Info.from_argparse(args)
-    # info is Dict, with 9 keys
-    assert len(info) == 9
+    # info is Dict, with 8 keys
+    assert len(info) == 8
 
-    # Three files are tifs
-    assert info.dst_gridinfo == 'input/directory/wrf_file_gridinfo.tif'
+    # Two files are tifs
     assert info.src_file == 'input/directory/lcz_file.tif'
     assert info.src_file_clean == 'input/directory/lcz_file_clean.tif'
 
@@ -111,7 +109,7 @@ def test_get_lcz_band_lcz_generator(info_mock, capsys):
         pytest.param(['--lcz-band', '5'], 5, id='custom band'),
     ),
 )
-def test_main_lcz_band_arg(arg, exp, info_mock, capsys, tmpdir):
+def test_main_lcz_band_arg(arg, exp, capsys, tmpdir):
     input_dir = tmpdir.mkdir('input')
     input_files = (
         'geo_em.d01.nc',
@@ -254,7 +252,6 @@ def test_check_lcz_wrf_extent_lcz_too_small(capsys, info_mock):
 
     out, _ = capsys.readouterr()
     assert ('ERROR: LCZ domain should be larger than WRF domain') in out
-    # TODO maybe add the actual values to check they are correct
     assert (
         'ERROR: LCZ domain should be larger than WRF domain '
         'in all directions.\nLCZ bounds (xmin, ymin, xmax, ymax): '
@@ -357,24 +354,25 @@ def test_wrf_remove_urban_output_already_exists_is_overwritten(tmpdir, info_mock
     assert m_time_old != os.path.getmtime(info.dst_nu_file)
 
 
-def test_create_wrf_gridinfo(tmpdir, info_mock):
+def test_get_wrf_grid_info(info_mock):
     info = info_mock(
         {
-            'dst_nu_file': 'testing/5by5.nc',
-            'dst_gridinfo': os.path.join(tmpdir, 'dst_gridinfo.tif'),
+            'dst_file': 'sample_data/geo_em.d04.nc',
         }
     )
-    create_wrf_gridinfo(info)
[…]
