+++ CITATION.cff
@@ -23,7 +23,7 @@ authors:
     email: jonas.kittner@rub.de
     affiliation: Urban Climatology Group, Department of Geography, Ruhr-University Bochum, Bochum, Germany
     orcid: 0000-0003-1881-7716
-version: v0.2.0
+version: v0.3.0
 date-released: 26-03-2022
 license: MIT
 repository-code: 'https://github.com/matthiasdemuzere/w2w'
