+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = w2w
-version = v0.2.0
+version = 0.3.0
 description = A python tool that ingests WUDAPT's Local Climate Zone information into WRF.
 long_description = file: README.md
 long_description_content_type = text/markdown
