### 2022-06-17 matthiasdemuzere/w2w#64 @thurber
[joss review] feedback
Hello `w2w` team, I am one of the reviewers for your JOSS submission tracked [here](https://github.com/openjournals/joss-reviews/issues/4432). I will leave my feedback in this issue. First of all, nice work! I had no problem installing the package or running the sample.

**Software**
- [ ] I wanted to test the other sample geo_em.d0X.nc files, but i was lazy and tried using the whole of Europe LCZ geotiff as input to `w2w`. This churned for awhile before failing and reporting a memory leak (see below). Not saying you should support this workflow, just that you might want to more explicitly warn about using a reasonably targeted LCZ raster.

```
> Processing MH_URB2D ...
[1]    59619 killed     w2w ./sample_data EU_LCZ_map.tif geo_em.d03.nc
/Users/thur961/.pyenv/versions/3.9.5/lib/python3.9/multiprocessing/resource_tracker.py:216: UserWarning: resource_tracker: There appear to be 1 leaked semaphore objects to clean up at shutdown
  warnings.warn('resource_tracker: There appear to be %d '
```

- [ ] Just as an FYI, I tried to reproduce the error from #63, but this worked correctly on my system
- [ ] There is a note that is output at the end of running `w2w` (see below). Is the user supposed to act on this? If not, may be best to suppress this or log it at a debug level instead.

```
 ----------- !! NOTE !! ---------
 Set nbui_max to 9 during compilation, in order to optimize memory storage.
```

**Paper**
- [ ] It took me a minute to realize some of the parenthetical content represented arguments to `w2w`, i.e. "(-f)", "(-l = 1)" and "(-b)" -- i like that this information is in the paper, but it could be nice to more explicitly call out the meaning and/or use the long form expression i.e. "(corresponding to parameter _--lcz-band_ = 1)"
- [ ] In the *Workflow* section, the indentation of each step is a bit awkward -- i recommend to use a small header tag or bold text instead

I'll give the paper another more in-depth read next week, and may have some more feedback at that point!
