### 2022-06-21 83db95d
respond to JOSS reviewer thurber
