### 2022-06-21 matthiasdemuzere/w2w#64 @matthiasdemuzere
Hi @thurber,

We really appreciated your first round of feedback on our work.
I just created a new branch, and pushed the recommended changes in there.
I'll leave this open for now, in case you have more suggestions after your second read.

Thanks again!
