### 2022-06-21 matthiasdemuzere/w2w#64 @matthiasdemuzere
I'll leave the compiled reviewed .pdf here: [j1iodiq3tayelfus2tlc.pdf](https://github.com/matthiasdemuzere/w2w/files/8946839/j1iodiq3tayelfus2tlc.pdf)
