### 2022-06-21 3acbc25
respond to JOSS reviewer thurber (#65)
* respond to JOSS reviewer thurber

* fix double dash?

* fix double dash?

* fix double dash?

* small typo
