### 2022-06-29 matthiasdemuzere/w2w#64 @thurber
Thanks @matthiasdemuzere! A few more comments and then I'm ready to approve:

**Software**
- [ ] Can you add a brief note to the README about how to run the automated tests? This can be as simple as "To Launch the automated tests, install pytest with `pip install pytest` and then run `pytest`.
- [ ] I'm getting a couple deprecation warnings when running the tests; it'd be a good idea to address those now to avoid future conflict

**Paper**
- [ ] I think opening the paper with a statement about WUDAPT gives a false impression that the new contribution is WUDAPT. Working in a sentence or two about w2w prior to the current opening would alleviate this.
- [ ] "more simple" --> "simpler"
- [ ] "netcdf" --> "NetCDF"
