### 2022-06-29 8460a93
add instructions on running the tests to readme, combine test coverage
