+++ tox.ini
@@ -7,6 +7,7 @@ deps = -rrequirements-dev.txt
 commands =
     coverage erase
     coverage run -m pytest {posargs:tests}
+    coverage combine
     coverage report
 
 
