### 2022-06-30 2f97c7a
add instructions on running the tests to readme, combine test coverage (#67)
