+++ setup.cfg
@@ -48,6 +48,7 @@ w2w.resources = LCZ_UCP_lookup.csv
 universal = True
 
 [coverage:run]
+parallel = True
 plugins = covdefaults
 
 [coverage:report]
