+++ w2w/w2w.py
@@ -415,6 +415,7 @@ def wrf_remove_urban(
     print(len(dst_data.south_north))
     for i in dst_data.south_north:
         from datetime import datetime
+
         startTime = datetime.now()
         print('i', i)
         for j in dst_data.west_east:
