### 2022-06-30 matthiasdemuzere/w2w#64 @matthiasdemuzere
Thanks @thurber  for the second round of comments.

I addressed the three paper items, and adjusted them according to your suggestions.
