### 2022-07-06 4306a8c
Improved efficiency wrf_remove_urban
