### 2022-07-08 16fef56
Update w2w.py
Correct selection of tiles when near borders
