### 2022-07-08 9c83799
Remove time prints
