+++ setup.cfg
@@ -22,6 +22,7 @@ classifiers =
 [options]
 packages = find:
 install_requires =
+    tqdm
     ecmwflibs
     numpy>=1.21
     pandas
