+++ setup.cfg
@@ -22,13 +22,13 @@ classifiers =
 [options]
 packages = find:
 install_requires =
-    tqdm
     ecmwflibs
     numpy>=1.21
     pandas
     rasterio>=1.1.6
     rioxarray
     scipy
+    tqdm
     xarray[io]
     importlib-metadata;python_version<"3.9"
     importlib-resources;python_version<"3.9"
