### 2022-07-11 65fe470
Optimize wrf remove urban
Using kdtree instead of brute force to search for N nearest neighbours to replace urban pixels with surrounding natural landscape. Added option to specify area size where the nearest grid points are searched (for optimization). Adapted tests. Removed ununsed functions.
