+++ tests/w2w_test.py
@@ -15,7 +15,6 @@ from rasterio.warp import Resampling
 
 import w2w.w2w
 from w2w.w2w import _add_frc_lu_index_2_wrf
-from w2w.w2w import _calc_distance_coord
 from w2w.w2w import _check_hi_values
 from w2w.w2w import _check_lcz_wrf_extent
 from w2w.w2w import _compute_hi_distribution
@@ -288,21 +287,6 @@ def test_check_lcz_integrity_clean_file_written(tmpdir, info_mock):
     check_lcz_integrity(info=info, LCZ_BAND=LCZ_BAND)
     assert os.listdir(tmpdir) == ['Shanghai_clean.tif']
 
-
-@pytest.mark.parametrize(
-    ('coords', 'expected'),
-    (
-        pytest.param((89, 0, 90, 1), 111194, id='near northern pole'),
-        pytest.param((-89, 0, -90, 1), 111194, id='near southern pole'),
-        pytest.param((-60, -10, -70, -12), 1115764, id='southern eastern hemisphere'),
-        pytest.param((-1, 0, 1, 2), 314498, id='across equator'),
-        pytest.param((45, 179, 46, -179), 191461, id='across dateline'),
-    ),
-)
-def test_calc_distance_coord(coords, expected):
-    assert _calc_distance_coord(*coords) == pytest.approx(expected, abs=1)
-
-
 @pytest.mark.parametrize(
     ('dst_file', 'dst_nu_file'),
     (
@@ -315,7 +299,7 @@ def test_wrf_remove_urban(tmpdir, dst_file, dst_nu_file, info_mock):
         {'dst_file': dst_file, 'dst_nu_file': os.path.join(tmpdir, dst_nu_file)}
     )
     old_ds = xr.open_dataset(info.dst_file)
-    wrf_remove_urban(info=info, NPIX_NLC=9)
+    wrf_remove_urban(info=info, NPIX_NLC=9,NPIX_AREA=25)
     ds = xr.open_dataset(info.dst_nu_file)
     # check lused 13 was reclassified to 12
     assert ds.LU_INDEX.values[0][2][2] == 12
@@ -350,7 +334,7 @@ def test_wrf_remove_urban_output_already_exists_is_overwritten(tmpdir, info_mock
     )
     assert os.listdir(tmpdir) == ['5by5_new.nc']
     m_time_old = os.path.getmtime(info.dst_nu_file)
-    wrf_remove_urban(info=info, NPIX_NLC=9)
+    wrf_remove_urban(info=info, NPIX_NLC=9,NPIX_AREA=25)
     assert m_time_old != os.path.getmtime(info.dst_nu_file)
 
 
