+++ w2w/w2w.py
@@ -8,6 +8,7 @@ import xarray as xr
 from rasterio.warp import reproject, Resampling
 from rasterio.transform import Affine
 from scipy.stats import mode, truncnorm
+import scipy.spatial as spatial
 import os, sys
 import argparse
 from argparse import RawTextHelpFormatter
@@ -113,11 +114,23 @@ def main(argv: Optional[Sequence[str]] = None) -> int:
         'natural land cover (DEFAULT: 45)',
         default=45,
     )
+
+    parser.add_argument(
+        '-a',
+        '--npix-area',
+        metavar='',
+        type=int,
+        dest='NPIX_AREA',
+        help='Area in number of pixels to look for the nearest number of pixels'
+        'for sampling neighbouring natural land cover (DEFAULT: NPIX_NLC**2)',
+        default=None,
+    )
     parser.add_argument(
         '--lcz-ucp',
         type=str,
         help='File with custom LCZ-based urban canopy parameters',
     )
+
     args = parser.parse_args(argv)
 
     # check if a custom LCZ UCP file was set and read it
@@ -127,7 +140,7 @@ def main(argv: Optional[Sequence[str]] = None) -> int:
         lookup_table = importlib_resources.files(
             'w2w.resources',
         ).joinpath('LCZ_UCP_lookup.csv')
-
+ 
     # Aesthetics, main prints in bold
     FBOLD = '\033[1m'
     FEND = '\033[0m'
@@ -150,9 +163,11 @@ def main(argv: Optional[Sequence[str]] = None) -> int:
     print(
         f'{FBOLD}--> Replace WRF MODIS urban LC with ' f'surrounding natural LC{FEND}'
     )
+
     wrf_remove_urban(
         info=info,
         NPIX_NLC=args.NPIX_NLC,
+        NPIX_AREA=args.NPIX_AREA,
     )
 
     print(f'{FBOLD}--> Create LCZ-based geo_em file{FEND}')
@@ -366,34 +381,32 @@ def check_lcz_integrity(info: Info, LCZ_BAND: int) -> None:
     lcz.rio.to_raster(info.src_file_clean, dtype=np.int32)
 
 
-def _calc_distance_coord(
-    lat1: float, lon1: float, lat2: float, lon2: float
-) -> xr.DataArray:
-
-    '''Calculate distance using coordinates
-    This uses the spherical law of cosines
-    '''
-
-    earth_radius = 6371000  # Earth radius in m
-    lat1r = lat1 * np.pi / 180.0
-    lat2r = lat2 * np.pi / 180.0
-    lon1r = lon1 * np.pi / 180.0
-    lon2r = lon2 * np.pi / 180.0
-
-    d = (
-        np.arccos(
-            np.sin(lat1r) * np.sin(lat2r)
-            + np.cos(lat1r) * np.cos(lat2r) * np.cos(lon2r - lon1r)
-        )
-        * earth_radius
-    )
-
-    return d
+def using_kdtree(data,kpoints):
+    "Based on https://stackoverflow.com/q/43020919/190597"
+    def dist_to_arclength(chord_length):
+        """
+        https://en.wikipedia.org/wiki/Great-circle_distance
+        Convert Euclidean chord length to great circle arc length
+        """
+        central_angle = 2*np.arcsin(chord_length/(2.0*R)) 
+        arclength = R*central_angle
+        return arclength
+    R = 6371
+    phi = np.deg2rad(data['lat'])
+    theta = np.deg2rad(data['lon'])
+    data['x'] = R * np.cos(phi) * np.cos(theta)
+    data['y'] = R * np.cos(phi) * np.sin(theta)
+    data['z'] = R * np.sin(phi)
[…]
