+++ w2w/w2w.py
@@ -140,7 +140,7 @@ def main(argv: Optional[Sequence[str]] = None) -> int:
         lookup_table = importlib_resources.files(
             'w2w.resources',
         ).joinpath('LCZ_UCP_lookup.csv')
- 
+
     # Aesthetics, main prints in bold
     FBOLD = '\033[1m'
     FEND = '\033[0m'
@@ -381,25 +381,27 @@ def check_lcz_integrity(info: Info, LCZ_BAND: int) -> None:
     lcz.rio.to_raster(info.src_file_clean, dtype=np.int32)
 
 
-def using_kdtree(data,kpoints):
-    "Based on https://stackoverflow.com/q/43020919/190597"
+def using_kdtree(data, kpoints):
+    'Based on https://stackoverflow.com/q/43020919/190597'
+
     def dist_to_arclength(chord_length):
         """
         https://en.wikipedia.org/wiki/Great-circle_distance
         Convert Euclidean chord length to great circle arc length
         """
-        central_angle = 2*np.arcsin(chord_length/(2.0*R)) 
-        arclength = R*central_angle
+        central_angle = 2 * np.arcsin(chord_length / (2.0 * R))
+        arclength = R * central_angle
         return arclength
+
     R = 6371
     phi = np.deg2rad(data['lat'])
     theta = np.deg2rad(data['lon'])
     data['x'] = R * np.cos(phi) * np.cos(theta)
     data['y'] = R * np.cos(phi) * np.sin(theta)
     data['z'] = R * np.sin(phi)
-    tree = spatial.KDTree(data[['x', 'y','z']])
-    distance, index = tree.query(data[['x', 'y','z']], k=kpoints)
-    #return dist_to_arclength(distance[:, 1])
+    tree = spatial.KDTree(data[['x', 'y', 'z']])
+    distance, index = tree.query(data[['x', 'y', 'z']], k=kpoints)
+    # return dist_to_arclength(distance[:, 1])
     return distance, index
 
 
@@ -425,70 +427,84 @@ def wrf_remove_urban(
     newgreenf = greenf.values.copy()
     orig_num_land_cat = dst_data.NUM_LAND_CAT
 
-    data_coord = pd.DataFrame({'lat':lat.values.ravel(), 'lon':lon.values.ravel(),'luse':luse.values.ravel()})
+    data_coord = pd.DataFrame(
+        {
+            'lat': lat.values.ravel(),
+            'lon': lon.values.ravel(),
+            'luse': luse.values.ravel(),
+        }
+    )
 
     ERROR = '\033[0;31m'
     ENDC = '\033[0m'
 
-  
-    luf_2D = luf.values.reshape(luf.shape[0],-1)
+    luf_2D = luf.values.reshape(luf.shape[0], -1)
     if orig_num_land_cat > 20:  # USING MODIS_LAKE
-        data_coord['luf_natland']=list((luf_2D[12,:]==0) & (luf_2D[16,:]==0) & (luf_2D[20,:]==0))
-        data_coord['luf_urb']=list((luf_2D[12,:]!=0))
+        data_coord['luf_natland'] = list(
+            (luf_2D[12, :] == 0) & (luf_2D[16, :] == 0) & (luf_2D[20, :] == 0)
+        )
+        data_coord['luf_urb'] = list((luf_2D[12, :] != 0))
     else:  # USING MODIS (NO LAKES)
-        data_coord['luf_natland']=list((luf_2D[12,:]==0) & (luf_2D[16,:]==0))
-        data_coord['luf_urb']=list((luf_2D[12,:]!=0))
-    if NPIX_AREA == None: NPIX_AREA=NPIX_NLC**2
-    if NPIX_AREA>luse.size:
-        raise ValueError(f'{ERROR}ERROR: The area you selected is larger than the domain size\n'
[…]
