### 2022-07-12 b53aa6a
Removed unnecessary function and added annotation
