### 2022-07-12 5a01850
add type annotation
