### 2022-07-12 2faec70
Update w2w.py
Set dkd to _, because we don't access it.
Added some comments
