### 2022-07-12 f869207
Update README.md
Added documentation of searching area for NPIX_NLC
