### 2022-07-13 a3b4c56
fix typo, simplify kd tree variable unpacking
