### 2022-07-13 61aa5f2
make w2w work on py310
