+++ setup.cfg
@@ -23,13 +23,15 @@ classifiers =
 packages = find:
 install_requires =
     ecmwflibs
+    h5netcdf
+    netCDF4
     numpy>=1.21
     pandas
     rasterio>=1.1.6
     rioxarray
     scipy
     tqdm
-    xarray[io]
+    xarray
     importlib-metadata;python_version<"3.9"
     importlib-resources;python_version<"3.9"
 python_requires = >=3.7
