### 2022-07-13 7dbb6a0
update JOSS with readme info
