### 2022-07-13 ecfc3ad
fix import sorting by moving np.seterr down
