### 2022-07-14 ebe9188
Update JOSS/paper.md
Co-authored-by: Jonas Kittner <54631600+theendlessriver13@users.noreply.github.com>
