### 2022-07-14 8e9e85d
Improved efficiency wrf_remove_urban (#71)
* Improved efficiency wrf_remove_urban

* Correct selection of tiles when near borders

* add tqdm for progress bar monitoring?

* Optimize wrf remove urban

* Using kdtree instead of brute force to search for N nearest neighbours to replace urban pixels with surrounding natural landscape. Added option to specify area size where the nearest grid points are searched (for optimization). Adapted tests. Removed ununsed functions.

* Removed unnecessary function and added annotation

* add type annotation

* Set dkd to _, because we don't access it. Added 
[…]
