### 2022-07-19 matthiasdemuzere/w2w#74 @erexer
[joss review] feedback 
Howdy, I'm a JOSS reviewer for your [submission](https://github.com/openjournals/joss-reviews/issues/4432) and I'll leave my initial feedback here. The example is running smoothly!

- [ ] `tox` tests all pass but you should address the warning it throws: `DeprecationWarning: SelectableGroups dict interface is deprecated. Use select.
    entrypoints = entry_points().get("xarray.backends", ())`
- [ ] If you enter an incorrect path for any inputs (ie. `w2w ./sample_data lcz_zaragoza.tif dne.nc`) you get a generic `FileNotFoundError` or `rasterio.errors.RasterioIOError`. It'd be nicer to have a custom error (ie. `cannot find GEO_EM file <input file> in <sample data dir>`).
- [ ] Might not be an issue, but it looks as if [one of your tests](https://github.com/matthiasdemuzere/w2w/blob/main/tests/w2w_test.py#L174) is looking at the `sample_data/lcz_zaragoza.tif` file. If this file is being modified in place, the tests shouldn't rely on it.

- [ ] It's not immediately clear that `w2w` modifies the `.nc` input file. Clarifying this in the `README` would be good.
- [ ] In [Important Notes](https://github.com/matthiasdemuzere/w2w#important-notes) the "chord-length formula" isn't hyperlinked to Wikipedia.
- [ ] It'd be nice to have some way for the user to confirm the functionality of the tutorial. Some sort of comparison of `geo_em.d04.nc` before and after running `w2w` would be cool.
- [ ] Again, may not be an issue; if the naming scheme `d0X` is ever going to be `d1X`, it'd be clearer to write it as `dX`, where `X` is a two-digit number.

#### Paper
- [ ] In the Statement of Need, you say that "make the translation of LCZ-based parameters better and more simple". It'd be good to have a concrete statement of why this is better and simpler than the Fortran package by Brousse et al.

I'll submit a PR with minor grammar for the paper but it's looking good!
