### 2022-07-20 8e1ac14
Fixed error message color
