### 2022-07-20 matthiasdemuzere/w2w#74 @dargueso
I made this plot to address comment 3 in document section.
@theendlessriver13 what do you think?
![W2Wchanges](https://user-images.githubusercontent.com/4198115/180025493-88d522e2-40bd-49d7-a349-21318448e4ee.png)
