### 2022-07-20 matthiasdemuzere/w2w#74 @dargueso
We should probably add a caption explaining. This is for the sample data.
