### 2022-07-20 matthiasdemuzere/w2w#74 @theendlessriver13
thanks this is looking awesome. I don't think this plotting routine should become part of w2w itself, but I think we should add this as an example python script, maybe to the sample_data folder? How are the plot colors chosen? We should probably check if they are accessible for color blind people.

This image could also probably be included in the `README` (with a link to the script), maybe at the bottom of the `Run the tool` section, having some visuals right away is always nice.
