### 2022-07-20 matthiasdemuzere/w2w#74 @dargueso
Hi Emily,

Thanks for the feedback!

Regarding comments 3 in **code** and 1 in **documentation**, `w2w` does not modify any files in place. It generates new files so the original ones remain untouched by the code. Not sure this answers your comment.

We can indicate that explicitly in the documentation to make it clearer.

We'll keep working on the other issues.

Thanks!
