### 2022-07-20 matthiasdemuzere/w2w#74 @erexer
> Regarding comments 3 in **code** and 1 in **documentation**, `w2w` does not modify any files in place. It generates new files so the original ones remain untouched by the code. Not sure this answers your comment.
>
> We can indicate that explicitly in the documentation to make it clearer.

gotcha, that makes a lot of sense and I see in the paper you state exactly what you output. I'll mark those as done since they're moot. It'd be good to state at least the final output file name in the README.
