### 2022-07-20 matthiasdemuzere/w2w#74 @theendlessriver13
yes, that's correct. I removed the stale branches now.
