### 2022-07-21 c8860d5
plot W2W results
Added script and image to visualize before and after w2w.
Added information to README on outputs and this figure
