+++ README.md
@@ -29,23 +29,23 @@ w2w ./sample_data lcz_zaragoza.tif geo_em.d04.nc
 
 3. Deploy using your own data:
 ```sh
-w2w INPUT_DIRECTORY YOUR_LCZ.TIF YOUR_GEO_EM.d0X.NC
+w2w INPUT_DIRECTORY YOUR_LCZ.TIF YOUR_GEO_EM.dXX.NC
 ```
 Output
 -------
 
    The **original files are not modified**. Three new files are generated by `w2w`:
    - geo_em.d04_NoUrban.nc: Urban areas are replaced with the dominant surrounding rural land use.
-   - geo_em.d04_LCZ_extent.nc: The extension of urban areas is updated using LCZ. 
-    - geo_em.d04_LCZ_params.nc: Urban areas are described using three different urban density categories. 
+   - geo_em.d04_LCZ_extent.nc: The extension of urban areas is updated using LCZ.
+    - geo_em.d04_LCZ_params.nc: Urban areas are described using three different urban density categories.
 
-   Differences between the original file and the newly created files by `w2w` are shown below for the sample data provided. 
+   Differences between the original file and the newly created files by `w2w` are shown below for the sample data provided.
 
    <p>
 <img src = "./sample_data/W2W-BeforeAfter.png">
 </p>
 <p>
-Fig.1: W2W results on sample data. Land use categories from the original file (a), the params file (b), the extent file (d) and the NoUrban (f). Pixels replaced with respect to the original file in params (c), extent (e) and NoUrb (g). 
+Fig.1: W2W results on sample data. Land use categories from the original file (a), the params file (b), the extent file (d) and the NoUrban (f). Pixels replaced with respect to the original file in params (c), extent (e) and NoUrb (g).
 </p>
 
    A [sample Jupyter notebook](./sample_data/Plotting_W2W-BeforeAfter.ipynb) is provided to recreate this Figure but is not part of the package, thus additional modules are required to run it.
@@ -53,7 +53,7 @@ Fig.1: W2W results on sample data. Land use categories from the original file (a
 
 Requirements
 -------
-1. A **geo_em.d0X.nc file** (produced by WRF's WPS geoegrid.exe), for the inner WRF model domain in which you would like to use the LCZ-based information.
+1. A **geo_em.dXX.nc file** (produced by WRF's WPS geoegrid.exe), for the inner WRF model domain in which you would like to use the LCZ-based information.
 
 
 2. A **Local Climate Zone map** (lcz.tif) that is slightly bigger than the domain of the geo_em.d0**X**.nc file. There are a number of ways to obtain an LCZ map for your region of interest:
@@ -65,11 +65,11 @@ Requirements
 
 Important notes
 -------
-* Your LCZ .tif and geo_em*.d0X.nc files should both live in the INPUT_DIRECTORY.
+* Your LCZ .tif and geo_em*.dXX.nc files should both live in the INPUT_DIRECTORY.
 * Also, this INPUT_DIRECTORY should be writeable by the user.
 * Also the geo_em.d0[**0 to X**].nc file(s) of the parent domain(s) should be available in the INPUT_DIRECTORY. This is needed because the `w2w.py` routine will check whether `NUM_LAND_CAT` is set to 41 in all these parent domain files. If that is not the case, this will be fixed by writing out adjusted geo_em.d0[**0 to X**]_41.nc files.
[…]
