+++ README.md
@@ -41,10 +41,7 @@ Output
 
    Differences between the original file and the newly created files by `w2w` are shown below for the sample data provided.
 
-   <p>
-<img src = "./sample_data/W2W-BeforeAfter.png">
-</p>
-<p>
+![before and after](./sample_data/W2W-BeforeAfter.png)
 Fig.1: w2w results on sample data. Land use categories from the original file (a), the params file (b), the extent file (d) and the NoUrban (f). Pixels replaced with respect to the original file in params (c), extent (e) and NoUrb (g).
 </p>
 
