### 2022-07-21 6b0bc7e
adapt pre-commit to work with ipynb files
