### 2022-07-22 75f12d5
set top_labels and right_labels as attribute
