### 2022-07-26 af9bb22
fix minor grammar and formatting
