+++ JOSS/paper.md
@@ -6,7 +6,7 @@ tags:
   - WRF
   - Local Climate Zones
   - Urban canopy parameters
-  - Climate modelling
+  - Climate modeling
 authors:
     - name: Matthias Demuzere^[corresponding author]
       orcid: 0000-0003-3237-4077
@@ -32,46 +32,46 @@ bibliography: paper.bib
 ---
 
 # Summary
-The Python-based WUDAPT-to-WRF (`W2W`) package is developed to translate Local Climate Zone (LCZ) maps into urban canopy parameters readable by WRF, the community "Weather Research and Forecasting" model [@Skamarock2021]. It is the successor of the Fortran-based `W2W` package developed by @Brousse2016 and @Martilli2016, and provides an improved, simpler, and more efficient procedure to use LCZ information in WRF. Some important changes include a direct manipulation of the geogrid files (without the creation of temporary files), and the use of average LCZ-based urban morphological parameters instead of assigning them to the modal LCZ class.
+The Python-based WUDAPT-to-WRF (`W2W`) package is developed to translate Local Climate Zone (LCZ) maps into urban canopy parameters readable by WRF, the community "Weather Research and Forecasting" model [@Skamarock2021]. It is the successor of the Fortran-based `W2W` package developed by @Brousse2016 and @Martilli2016, and provides an improved, simpler, and more efficient procedure to use LCZ information in WRF. Some important changes include direct manipulation of the geogrid files without the creation of temporary files, and the use of average LCZ-based urban morphological parameters instead of assigning them to the modal LCZ class.
 
-This development of this package is in line with the objectives of WUDAPT, the World Urban Database and Access Portals Tools community project, that aims to 1) acquire and make accessible coherent and consistent information on form and function of urban morphology relevant to climate weather, and environment studies, and 2) provide tools that extract relevant urban parameters and properties for models and model applications at appropriate scales for various climate, weather, environment, and urban planning purposes [@Ching2018].
+This development of this package is in line with the objectives of WUDAPT, the World Urban Database and Access Portals Tools community project, that aims to 1) acquire and make accessible coherent and consistent information on the form and function of urban morphology relevant to climate weather and environmental studies, and 2) provide tools that extract relevant urban parameters and properties for models and model applications at appropriate scales for various climate, weather, environment, and urban planning purposes [@Ching2018].
 
 # Statement of need
 Since the pioneering work of @Brousse2016 and @Martilli2016, the level-0 WUDAPT information, the Local Climate Zone maps, have been used increasingly in WRF.
 
-We expect this trend to continue, because of three recent developments: 1) the creation of city-wide LCZ maps is now easier than ever with the launch of the LCZ Generator web application [@Demuzere2021], 2) the availability of a global LCZ map [@Demuzere2022], and 3) WRF versions > 4.3 [@Skamarock2021] are able to ingest 10 or 11 built classes (corresponding to WUDAPT's LCZs) by default, whereas previous WRF versions required manual code changes (see @Martilli2016, @Zonato2021a and @Zonato2021b for more information).
[…]
