### 2022-08-08 matthiasdemuzere/w2w#74 @matthiasdemuzere
Thanks for all the work @dargueso and @theendlessriver13.

I can address the last point. Just to be sure: what branch are we currently working in?

@dargueso: for the figure you point to a Jupyter notebook, yet the link provided in README does not seem to work? This one: [https://github.com/matthiasdemuzere/w2w/blob/main/sample_data/Plotting_W2W-BeforeAfter.ipynb](https://github.com/matthiasdemuzere/w2w/blob/main/sample_data/Plotting_W2W-BeforeAfter.ipynb)

Ow, wait, I see it is just a filename issue, this seems to be the correct link: [https://github.com/matthiasdemuzere/w2w/blob/main/sample_data/plotting_w2w_before_after.ipynb](https://github.com/matthiasdemuzere/w2w/blob/main/sample_data/plotting_w2w_before_after.ipynb)

I can fix this ....
