+++ README.md
@@ -45,7 +45,7 @@ Output
 Fig.1: w2w results on sample data. Land use categories from the original file (a), the params file (b), the extent file (d) and the NoUrban (f). Pixels replaced with respect to the original file in params (c), extent (e) and NoUrb (g).
 </p>
 
-   A [sample Jupyter notebook](./sample_data/Plotting_W2W-BeforeAfter.ipynb) is provided to recreate this Figure but is not part of the package, thus additional modules are required to run it.
+   A [sample Jupyter notebook](./sample_data/plotting_w2w_before_after.ipynb) is provided to recreate this Figure but is not part of the package, thus additional modules are required to run it.
 
 
 Requirements
