+++ JOSS/paper.bib
@@ -230,7 +230,7 @@
     author = {Martilli, Alberto and Brousse, Oscar and Ching, Jason},
     publisher = {http://www.wudapt.org/wudapt-to-wrf/},
     title = {{Urbanized WRF modeling using WUDAPT}},
-    url = {https://www.wudapt.org/wp-content/uploads/2016/05/Urbanized-WRF-modeling-using-WUDAPT-web-version-March2016.pdf (Accessed on 11 August 2021},
+    url = {https://www.wudapt.org/wp-content/uploads/2016/05/Urbanized-WRF-modeling-using-WUDAPT-web-version-March2016.pdf (Last accessed on 8 August 2022},
     year = {2016}
 }
 
