+++ JOSS/paper.md
@@ -41,7 +41,7 @@ Since the pioneering work of @Brousse2016 and @Martilli2016, the level-0 WUDAPT 
 
 We expect this trend to continue because of three recent developments: 1) the creation of city-wide LCZ maps is now easier than ever with the launch of the LCZ Generator web application [@Demuzere2021], 2) the availability of a global LCZ map [@Demuzere2022], and 3) WRF versions > 4.3 [@Skamarock2021] can ingest 10 or 11 built classes (corresponding to WUDAPT's LCZs) by default, whereas previous WRF versions required manual code changes (see @Martilli2016, @Zonato2021a and @Zonato2021b for more information).
 
-Because of these developments, an improved, Python-based, WUDAPT-to-WRF (`W2W`) routine is presented here to translate LCZ-based parameters better and simpler. It differs from its Fortran-based predecessor mainly by 1) using a more up-to-date LCZ-based urban extent, 2) the aggregation of the morphological parameters instead of using modal values, and 3) the fact that all processing is done with one (automated) tool, whereas the Fortran-based version required multiple pre-processing steps and manual namelist changes, described in more detail by [@Martilli2016].  
+Because of these developments, an improved, Python-based, WUDAPT-to-WRF (`W2W`) routine is presented here to translate LCZ-based parameters better and simpler. It differs from its Fortran-based predecessor mainly by 1) using a more up-to-date LCZ-based urban extent, 2) the aggregation of the morphological parameters instead of using modal values, and 3) the fact that all processing is done with one (automated) tool, whereas the Fortran-based version required multiple pre-processing steps and manual namelist changes, described in more detail by [@Martilli2016].
 
 # Initial data requirements
 To use the tool, two input files are required:
