### 2022-08-08 1bfe45b
address last JOSS review comment
