### 2022-08-08 #4432 @matthiasdemuzere
@crvernon I just addressed the last point (see [here](https://github.com/matthiasdemuzere/w2w/pull/83)), and I am awaiting approval from my colleagues.

Once in, I hope this sufficiently addresses all helpful feedback provided by @erexer!
