### 2022-08-08 90bec54
update .ipynb url, address last paper review (#83)
* update .ipynb url

* address last JOSS review comment
