### 2022-08-08 matthiasdemuzere/w2w#74 @matthiasdemuzere
Hi @erexer,

Thanks for the useful feedback.
I have now completed the last checkbox, I hope this is in line with what you had in mind!
