### 2022-08-15 #4432 @crvernon
@matthiasdemuzere here are a few outstanding things you need to address:

In `paper.md`:

- [ ] [LINE 73] need an actual citation for "the chord length formula" and not a link to a wikipedia page, which can change often.

Once the above change has been made:

- [ ]  Conduct a GitHub release of the current reviewed version of the software you now have on the main and archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository) - you can link your GitHub repository to automatically generate your DOI on release.  Zenodo will recognize your `CITATION.cff` file and populate the metadata of your release with that information.
- [ ]   Please list the DOI of the archived version here

I can then move forward with accepting the submission!
