### 2022-08-22 22dbf3b
Details on KDtree algorithm and distance
Clarified the calculation of distance and fixed references
