+++ JOSS/paper.bib
@@ -1,390 +1,408 @@
+%% This BibTeX bibliography file was created using BibDesk.
+%% https://bibdesk.sourceforge.io/
+
+%% Created for Daniel Argüeso at 2022-08-22 11:17:17 +0200 
+
+
+%% Saved with string encoding Unicode (UTF-8) 
+
+
+
+@article{Maneewongvatana1999,
+	author = {Maneewongvatana, Songrit and Mount, David M.},
+	date-added = {2022-08-22 11:12:42 +0200},
+	date-modified = {2022-08-22 11:16:26 +0200},
+	doi = {10.48550/arXiv.cs/9901013},
+	journal = {Computer Science - ArXiv},
+	title = {{Analysis of Approximate Nearest Neighbor Searching with Clustered Point Sets}},
+	year = {1999},
+	bdsk-url-1 = {https://essd.copernicus.org/preprints/essd-2022-92/},
+	bdsk-url-2 = {https://doi.org/10.5194/essd-2022-92}}
+
 @article{Demuzere2022,
-    abstract = {},
-    author = {Demuzere, Matthias and Kittner, Jonas and Martilli, Alberto and Mills, Gerald and Moede, Christian and Stewart, Iain D. and van Vliet, Jasper and Bechtel, Benjamin},
-    doi = {10.5194/essd-2022-92},
-    issn = {},
-    journal = {Earth System Science Data Discussions},
-    title = {{A global map of Local Climate Zones to support earth system modelling and urban scale environmental science}},
-    url = {https://essd.copernicus.org/preprints/essd-2022-92/},
-    volume = {},
-    year = {2022}
-}
+	author = {Demuzere, Matthias and Kittner, Jonas and Martilli, Alberto and Mills, Gerald and Moede, Christian and Stewart, Iain D. and van Vliet, Jasper and Bechtel, Benjamin},
+	doi = {10.5194/essd-2022-92},
+	journal = {Earth System Science Data Discussions},
+	title = {{A global map of Local Climate Zones to support earth system modelling and urban scale environmental science}},
+	url = {https://essd.copernicus.org/preprints/essd-2022-92/},
+	year = {2022},
+	bdsk-url-1 = {https://essd.copernicus.org/preprints/essd-2022-92/},
+	bdsk-url-2 = {https://doi.org/10.5194/essd-2022-92}}
 
 @misc{Glotfelty2013,
-    author = {Glotfelty, T. and Tewari, Mukul and Sampson, Kevin and Duda, Michael and Chen, Fei and Ching, Jason},
-    pages = {9},
-    publisher = {National Center for Atmospheric Research},
-    title = {{NUDAPT 44: How to use NUDAPT dataset in WRF/SLUCM/MLUCM models}},
-    url = {https://www.yumpu.com/en/document/read/26871494/how-to-use-nudapt-dataset-in-wrf-slucm-mlucm-models},
-    year = {2013}
-}
-
+	author = {Glotfelty, T. and Tewari, Mukul and Sampson, Kevin and Duda, Michael and Chen, Fei and Ching, Jason},
+	pages = {9},
+	publisher = {National Center for Atmospheric Research},
+	title = {{NUDAPT 44: How to use NUDAPT dataset in WRF/SLUCM/MLUCM models}},
+	url = {https://www.yumpu.com/en/document/read/26871494/how-to-use-nudapt-dataset-in-wrf-slucm-mlucm-models},
+	year = {2013},
+	bdsk-url-1 = {https://www.yumpu.com/en/document/read/26871494/how-to-use-nudapt-dataset-in-wrf-slucm-mlucm-models}}
 
 @article{Stewart2014,
-    author = {Stewart, {Iain D.} and Oke, {T. R.} and Krayenhoff, {E. Scott}},
[…]
