### 2022-08-22 f2b98db
Update README.md
Removed chord lenght formula, because the Euclidean distance over the great circle arch is used instead
