### 2022-08-22 #4432 @dargueso
We have removed references to "the chord length formula" because the actual distance over the great circle arch is used instead. This was specified in the description of the function and the corresponding links, but was not updated in the README.md and the manuscript itself.  We have better described the method, including a definition of the great circle arch, and a reference for the KDtree algorithm.
