### 2022-08-22 #4432 @dargueso
> @matthiasdemuzere here are a few outstanding things you need to address:
>
> In `paper.md`:
>
> * [ ]  [LINE 73] need an actual citation for "the chord length formula" and not a link to a wikipedia page, which can change often.
>

We have clarified the method and added references for the KDtree algorithm. We used the Euclidean distance as a metric of proximity over the 3D space by converting geographical coordinates (R, lat, lon) to cartesian coordinates (x, y ,z). The resulting length is the chord of the great circle arc. We added a definition of the great circle arc to the text, but we think the chord is a well established term and does not need a reference.

Hope this make the method clearer.
