### 2022-08-22 #4432 @crvernon
No problem @matthiasdemuzere and @dargueso

Once you get the following knocked out, we will move forward.  Thanks!

- [ ]  Conduct a GitHub release of the current reviewed version of the software you now have on the main and archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository) - you can link your GitHub repository to automatically generate your DOI on release.  Zenodo will recognize your `CITATION.cff` file and populate the metadata of your release with that information.
- [ ]   Please list the DOI of the archived version here
