+++ JOSS/paper.md
@@ -63,7 +63,7 @@ To get to that point, a number of sequential steps are required:
 
 ### Step 1: Remove the default urban land cover
 
-The default urban land cover from MODIS is replaced with the dominant surrounding vegetation category, as done in @Li2020. This procedure affects WRF's parameters LU_INDEX, LANDUSEF, and GREENFRAC. First, an initial number of neighboring pixels (corresponding argument `--npix-area`, default = `--npix-nlc` ** 2) are selected using scipy's k-d tree algorithm [@Maneewongvatana1999], assuming an spherical Earth, and using the Euclidean distance along the great circle arc (i.e., section of the Earth that divides the sphere into two equal parts) to find the nearest pixels. Afterward, the LU_INDEX is set by sampling the dominant category from the corresponding argument `--npix-nlc` (default = 45) nearest initial grid points (excluding ocean, urban, and lakes). GREENFRAC is calculated as the mean over all grid points with that dominant vegetation category among the `--npix-nlc` nearest points. For each grid point, if LANDUSEF had any percentage of urban, it is set to zero and the percentage is added to the dominant vegetation category assigned to that grid point.
+The default urban land cover from MODIS is replaced with the dominant surrounding vegetation category, as done in @Li2020. This procedure affects WRF's parameters LU_INDEX, LANDUSEF, and GREENFRAC. First, an initial number of neighboring pixels (corresponding argument `--npix-area`, default = `--npix-nlc` ** 2) are selected using scipy's k-d tree algorithm [@Maneewongvatana1999], assuming a spherical Earth, and using the Euclidean distance along the great circle arc (i.e., section of the Earth that divides the sphere into two equal parts) to find the nearest pixels. Afterwards, the LU_INDEX is set by sampling the dominant category from the corresponding argument `--npix-nlc` (default = 45) nearest initial grid points (excluding ocean, urban, and lakes). GREENFRAC is calculated as the mean over all grid points with that dominant vegetation category among the `--npix-nlc` nearest points. For each grid point, if LANDUSEF had any percentage of urban, it is set to zero and the percentage is added to the dominant vegetation category assigned to that grid point.
 
 Resulting output: **`geo_em.dXX_NoUrban.nc`**
 
