+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = w2w
-version = 0.4.0
+version = 0.4.1
 description = A python tool that ingests WUDAPT's Local Climate Zone information into WRF.
 long_description = file: README.md
 long_description_content_type = text/markdown
