### 2022-08-23 #4432 @matthiasdemuzere
> * [x]   Conduct a GitHub release of the current reviewed version of the software you now have on the main and archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository) - you can link your GitHub repository to automatically generate your DOI on release.  Zenodo will recognize your `CITATION.cff` file and populate the metadata of your release with that information.

I have released a new version 0.4.1, and created a Zenodo repo for this: [10.5281/zenodo.7016607](https://zenodo.org/record/7016607)

> * [x]    Check the archival deposit (e.g., in Zenodo) to ensure it has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.

I checked this things, looks fine to me.

> * [x]    Please list the DOI of the archived version here

10.5281/zenodo.7016607

I hope all of the above is as expected?
