### 2022-08-23 #4432 @crvernon
See previous note about missing DOI:
https://github.com/openjournals/joss-reviews/issues/4432#issuecomment-1214992476
